+++ src/MAUI/Maui.Samples/Samples/Layers/QueryRelatedFeaturesFromNonSpatialTable/QueryRelatedFeaturesFromNonSpatialTable.xaml.cs
@@ -16,20 +16,20 @@ using Esri.ArcGISRuntime.UI;
 
 using Color = System.Drawing.Color;
 
-namespace ArcGIS.Samples.MapImageLayerTables
+namespace ArcGIS.Samples.QueryRelatedFeaturesFromNonSpatialTable
 {
     [ArcGIS.Samples.Shared.Attributes.Sample(
-        name: "Map image layer tables",
+        name: "Query related features from non spatial table",
         category: "Layers",
         description: "Find features in a spatial table related to features in a non-spatial table.",
         instructions: "Once the map image layer loads, a list view will be populated with comment data from non-spatial features. Tap on one of the comments to query related spatial features and display the first result on the map.",
         tags: new[] { "features", "query", "related features", "search" })]
-    public partial class MapImageLayerTables : ContentPage
+    public partial class QueryRelatedFeaturesFromNonSpatialTable : ContentPage
     {
         // A graphics overlay for showing selected features.
         private GraphicsOverlay _selectedFeaturesOverlay;
 
-        public MapImageLayerTables()
+        public QueryRelatedFeaturesFromNonSpatialTable()
         {
             InitializeComponent();
