+++ src/WPF/WPF.Viewer/Samples/Layers/QueryRelatedFeaturesFromNonSpatialTable/QueryRelatedFeaturesFromNonSpatialTable.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGIS.WPF.Samples.MapImageLayerTables.MapImageLayerTables"
+﻿<UserControl x:Class="ArcGIS.WPF.Samples.QueryRelatedFeaturesFromNonSpatialTable.QueryRelatedFeaturesFromNonSpatialTable"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
