+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Layers/QueryRelatedFeaturesFromNonSpatialTable/QueryRelatedFeaturesFromNonSpatialTable.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGIS.WinUI.Samples.MapImageLayerTables.MapImageLayerTables"
+﻿<UserControl x:Class="ArcGIS.WinUI.Samples.QueryRelatedFeaturesFromNonSpatialTable.QueryRelatedFeaturesFromNonSpatialTable"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls">
