+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Layers/QueryRelatedFeaturesFromNonSpatialTable/readme.metadata.json
@@ -1,10 +1,10 @@
 {
     "category": "Layers",
     "description": "Find features in a spatial table related to features in a non-spatial table.",
-    "formal_name": "MapImageLayerTables",
+    "formal_name": "QueryRelatedFeaturesFromNonSpatialTable",
     "ignore": false,
     "images": [
-        "mapimagelayertables.jpg"
+        "QueryRelatedFeaturesFromNonSpatialTable.jpg"
     ],
     "keywords": [
         "features",
@@ -14,7 +14,7 @@
     ],
     "offline_data": [],
     "redirect_from": [
-        "/net/latest/maui/sample-code/map-image-layer-tables.htm"
+        "/net/latest/winui/sample-code/map-image-layer-tables.htm"
     ],
     "relevant_apis": [
         "ArcGISFeature",
@@ -28,8 +28,8 @@
         "ServiceFeatureTable"
     ],
     "snippets": [
-        "MapImageLayerTables.xaml.cs",
-        "MapImageLayerTables.xaml"
+        "QueryRelatedFeaturesFromNonSpatialTable.xaml.cs",
+        "QueryRelatedFeaturesFromNonSpatialTable.xaml"
     ],
-    "title": "Map image layer tables"
+    "title": "Query related features from non spatial table"
 }
